Remove retry logic and push directly in dead letter Retry logic added more complexity overall and is not very usefull. If an error occure during message processing either it's due to a stress situation of other systems, either it's a bug generated by this specific message. In both cases it's better to not retry to send the message, so we put it directly in a dead letter queue for later investigation. We leave the possibility to panic after a certain number of published dead lettered messages in order to quit if all messages are failing like in a stressed system scenario.

+++ evaldgraph/main.go
@@ -69,7 +69,7 @@ func run() error {
 		                                                       &ulh.UserLocationHandler {
 																														 DgCl: dgCl,
 																													 },
-		                                                       uint32(0),
+		                                                       0,
 		                                                       msgDr)
 
 	defer producingMsgCons.Close()
